+++ src/components/spartan/SInputIncrement/index.ts
@@ -1,1 +1,3 @@
 export { default as SInputIncrement } from './SInputIncrement.vue';
+
+export type { TInputIncrementProps, TInputIncrementEmits } from './types';